+++ R/TheiaCollection.R
@@ -14,8 +14,9 @@
 #'                             query     = NULL,
 #'                             dir.path  = NULL,
 #'                             check     = TRUE)
+#'                             quiet     = TRUE)
 #'
-#'    c$download(auth, overwrite = FALSE, check = TRUE)
+#'    c$download(auth, overwrite = FALSE, check = TRUE, quiet = TRUE)
 #'    c$check()
 #'    c$status
 #'    c$extract(overwrite = FALSE, dest.dir = NULL)
@@ -29,6 +30,7 @@
 #'    \item{c:}{A \code{TheiaCollection} object}
 #'    \item{dir.path:}{The path to the directory containing zip files}
 #'    \item{check:}{Whether or not to check existing files on collection's creation}
+#'    \item{quiet:}{Control verbose output}
 #'    \item{tiles:}{A list of TheiaTile objects}
 #'    \item{cart:}{An XML cart parsed from a 'meta4' file downloaded from Theia}
 #'    website. Used only if Collection is created from a cart
@@ -120,9 +122,16 @@ TheiaCollection <-
                                        tiles      = NULL,
                                        query      = NULL,
                                        dir.path   = NULL,
-                                       check      = TRUE)
+                                       check      = TRUE,
+                                       quiet      = TRUE)
                  {
-                   .TheiaCollection_initialize(self, cart.path, tiles, query, dir.path, check)
+                   if (quiet == TRUE) {
+                     suppressMessages({
+                       .TheiaCollection_initialize(self, cart.path, tiles, query, dir.path, check)
+                     })
+                   } else {
+                     .TheiaCollection_initialize(self, cart.path, tiles, query, dir.path, check)
+                   }
                  },
 
                  print = function(...)
@@ -135,9 +144,9 @@ TheiaCollection <-
                    .TheiaCollection_check(self)
                  },
 
-                 download = function(auth, overwrite = FALSE, check = TRUE)
+                 download = function(auth, overwrite = FALSE, check = TRUE, quiet = TRUE)
                  {
-                   .TheiaCollection_download(self, auth, overwrite, check)
+                   .TheiaCollection_download(self, auth, overwrite, check, quiet)
                  },
 
                  extract = function(overwrite = FALSE, dest.dir = NULL)
@@ -257,7 +266,7 @@ TheiaCollection <-
 }
 
 
-.TheiaCollection_download <- function(self, auth, overwrite, check)
+.TheiaCollection_download <- function(self, auth, overwrite, check, quiet)
 {
   if (is.character(auth)) {
     # create authentification system if not supplied
@@ -266,10 +275,10 @@ TheiaCollection <-
 
   # download needed tiles
   lapply(self$tiles,
-         function(x, auth, overwrite, check) {
-           x$download(auth, overwrite = overwrite, check = check)
+         function(x, auth, overwrite, check, quiet) {
+           x$download(auth, overwrite = overwrite, check = check, quiet)
          },
-         auth = auth, overwrite = overwrite, check = check)
+         auth = auth, overwrite = overwrite, check = check, quiet = quiet)
 
   return(invisible(self))
 }